# public/swagger/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "My First API",
        "version": "2.0"
    },
    "paths": {
        "/translator": {
            "post": {
                "summary": "Translate Text",
                "description": "Translate text from one language to another.",
                "operationId": "fd79ad4497de7df04502d824700f21fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "q": {
                                        "description": "Text to be translated.",
                                        "type": "string"
                                    },
                                    "source": {
                                        "description": "Source language code.",
                                        "type": "string"
                                    },
                                    "target": {
                                        "description": "Target language code.",
                                        "type": "string"
                                    },
                                    "format": {
                                        "description": "Translation format.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful translation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Translation successful"
                                        },
                                        "translated_text": {
                                            "type": "string",
                                            "example": "Translated text goes here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error in translation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Translation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/getapiusage": {
            "get": {
                "summary": "Get API Usage History",
                "description": "Retrieve the history of API usage (entire data in the history table).",
                "operationId": "d2d648326d624331d5323f8b17d61bb9",
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "API usage history retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },

        "/translatefile": {
            "post": {
              "summary": "Translate and save CSV file content",
              "description": "Translate the content of a CSV file and save it to the database.",
              "requestBody": {
                "required": true,
                "content": {
                  "multipart/form-data": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "csvfile": {
                          "type": "string",
                          "format": "binary"
                        }
                      }
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "CSV file translated and saved successfully",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "statusCode": {
                            "type": "integer",
                            "example": 200
                          },
                          "message": {
                            "type": "string",
                            "example": "CSV file translated and saved."
                          }
                        }
                      }
                    }
                  }
                },
                "422": {
                  "description": "Unprocessable Entity",
                  "content": {
                    "application/json": {
                      "schema": {
                        "type": "object",
                        "properties": {
                          "statusCode": {
                            "type": "integer",
                            "example": 422
                          },
                          "message": {
                            "type": "string",
                            "example": "Error message if applicable"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
          
    }
}